requirements.txt
amqp==1.4.9
anyjson==0.3.3
billiard==3.3.0.23
celery==3.1.23
dj-database-url==0.4.1
dj-static==0.0.6
Django==1.7
django-easymode==1.4b5
django-paypal==0.3.2
gunicorn==19.6.0
kombu==3.0.35
MySQL-python==1.2.5
Pillow==3.2.0
polib==1.0.7
pycrypto==2.6.1
pytz==2016.4
redis==2.10.5
requests==2.11.1
simple-crypt==4.1.7
six==1.10.0
static3==0.7.0
